requirements.txt
BeautifulSoup4

# Testing
pytest
pytest-cov
pytest-mock

